+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * [\#68](https://github.com/Finschia/wasmd/pull/68) add an integration test for UpdateAdmin
 * [\#99](https://github.com/Finschia/wasmd/pull/99) format test files
 * [\#98](https://github.com/Finschia/wasmd/pull/98) refactor TestStoreAndInstantiateContract
+* [\#100](https://github.com/Finschia/wasmd/pull/100) refactor tests for cosmwasm/wasm/v1/tx.proto other than TestClearAdmin
 
 ### Bug Fixes
 * [\#62](https://github.com/Finschia/wasmd/pull/62) fill ContractHistory querier result's Updated field